+++ DietScoopApp/app/src/main/AndroidManifest.xml
@@ -13,16 +13,16 @@
         android:theme="@style/Theme.DietScoop"
         tools:targetApi="31">
         <activity
-            android:name=".ViewRecipeActivity"
+            android:name=".Activities.ViewRecipeActivity"
             android:exported="false" />
         <activity
-            android:name=".RecipeListActivity"
+            android:name=".Activities.RecipeListActivity"
             android:exported="false" />
         <activity
-            android:name=".IngredientListActivity"
+            android:name=".Activities.IngredientListActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".Activities.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
