+++ DietScoopApp/app/src/main/java/com/example/dietscoop/Data/TypeIsNumeric.java
@@ -1,0 +1,6 @@
+package com.example.dietscoop.Data;
+
+public enum TypeIsNumeric {
+    integer,
+    decimal
+}
