+++ DietScoopApp/app/src/main/java/com/example/dietscoop/Fragments/sortIngredientByFragment.java
@@ -1,4 +1,4 @@
-package com.example.dietscoop;
+package com.example.dietscoop.Fragments;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
@@ -9,6 +9,8 @@ import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.Button;
+
+import com.example.dietscoop.R;
 
 /**
  * Fragment class for prompting user with sorting options for ingredients
