+++ DietScoopApp/app/src/main/res/layout/activity_ingredient_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".IngredientListActivity">
+    tools:context=".Activities.IngredientListActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/ingredient_list"
