+++ DietScoopApp/app/src/main/res/layout/activity_recipe_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".RecipeListActivity">
+    tools:context=".Activities.RecipeListActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recipe_list"
