+++ DietScoopApp/app/src/main/res/layout/activity_view_recipe.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ViewRecipeActivity">
+    tools:context=".Activities.ViewRecipeActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
